tests(vue-component): refactor to reuse common code
